Combine all reset password actions in one file
