+++ src/tests/sagas/postSetNewPasswordExtraSecurityTokenSaga-test.js
@@ -2,7 +2,7 @@ import expect from "expect";
 import { call } from "redux-saga/effects";
 import postRequest from "../../login/redux/sagas/postDataRequest";
 import { postSetNewPasswordExtraSecurityToken } from "../../login/redux/sagas/resetpassword/postSetNewPasswordExtraSecurityTokenSaga";
-import { setNewPasswordExtraSecurityTokenFail } from "../../login/redux/actions/postResetNewPasswordActions";
+import { setNewPasswordExtraSecurityTokenFail } from "../../login/redux/actions/postResetPasswordActions";
 import { safeEncode } from "../../login/app_utils/helperFunctions/authenticatorAssertion";
 
 const fakeState = {
